pyproject.toml
[tool.poetry]
name = "git-metrics"
version = "0.1.0"
description = ""
authors = ["Adam Blake <adamblake@g.ucla.edu>"]
license = "AGPL-3+"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.27"
click = "^8.1.3"
pathlib = "^1.0.1"
python-dateutil = "^2.8.2"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
flake8 = "^4.0.1"
mypy = "^0.971"

[tool.poetry.scripts]
similarity = "git_metrics.cli:similarity"
activity = "git_metrics.cli:activity"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
